Move InstanceMethods module into its own source file
